requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --output-file=requirements/base.txt --pip-args=None requirements/base.in
#
alembic==1.16.5
    # via -r requirements/base.in
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
click==8.3.0
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi[standard]==0.117.1
    # via -r requirements/base.in
fastapi-cli[standard]==0.0.13
    # via fastapi
fastapi-cloud-cli==0.2.0
    # via fastapi-cli
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.6
    # via fastapi
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
passlib==1.7.4
    # via -r requirements/base.in
psycopg[binary]==3.2.10
    # via -r requirements/base.in
psycopg-binary==3.2.10
    # via psycopg
pydantic[email]==2.11.9
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via -r requirements/base.in
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via -r requirements/base.in
python-dotenv==1.1.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
rich==14.1.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.6.4
    # via fastapi-cloud-cli
sentry-sdk==2.38.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   -r requirements/base.in
    #   alembic
starlette==0.48.0
    # via fastapi
typer==0.19.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via sentry-sdk
uvicorn[standard]==0.36.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
